+++ mininterface/tk_interface/date_entry.py
@@ -1,42 +1,66 @@
 import tkinter as tk
 import re
 from datetime import datetime
+from typing import TYPE_CHECKING
 
 try:
     from tkcalendar import Calendar
 except ImportError:
     Calendar = None
 
-class DateEntry(tk.Frame):
-    def __init__(self, master=None, **kwargs):
+from ..types import DatetimeTag
+if TYPE_CHECKING:
+    from tk_window import TkWindow
+
+
+class DateEntryFrame(tk.Frame):
+    def __init__(self, master, tk_app: "TkWindow", tag: DatetimeTag, variable: tk.Variable, **kwargs):
         super().__init__(master, **kwargs)
-        self.create_widgets()
-        self.pack(expand=True, fill=tk.BOTH)
+
+        self.tk_app = tk_app
+        self.tag = tag
+
+        # Date entry
+        self.spinbox = self.create_spinbox(variable)
+
+        # Frame holding the calendar
+        self.frame = tk.Frame(self)
+
+        # The calendar widget
+        if Calendar:
+            # Toggle calendar button
+            tk.Button(self, text="…", command=self.toggle_calendar).grid(row=0, column=1)
+
+            # Add a calendar widget
+            self.calendar = Calendar(self.frame, selectmode='day', date_pattern='yyyy-mm-dd')
+            # Bind date selection event
+            self.calendar.bind("<<CalendarSelected>>", self.on_date_select)
+            self.calendar.grid()
+            # Initialize calendar with the current date
+            self.update_calendar(self.spinbox.get(), '%Y-%m-%d %H:%M:%S.%f')
+        else:
+            self.calendar = None
+
         self.bind_all_events()
 
-    def create_widgets(self):
-        self.spinbox = tk.Spinbox(self, font=("Arial", 16), width=30, wrap=True)
-        self.spinbox.pack(padx=20, pady=20)
-        self.spinbox.insert(0, datetime.now().strftime("%Y-%m-%d %H:%M:%S.%f")[:-4])
-        self.spinbox.focus_set()
-        self.spinbox.icursor(8)
+    def create_spinbox(self, variable: tk.Variable):
+        spinbox = tk.Spinbox(self, font=("Arial", 16), width=30, wrap=True, textvariable=variable)
+        spinbox.grid()
+        if not variable.get():
+            spinbox.insert(0, datetime.now().strftime("%Y-%m-%d %H:%M:%S.%f")[:-4])
+        spinbox.focus_set()
+        spinbox.icursor(8)
 
         # Bind up/down arrow keys
-        self.spinbox.bind("<Up>", self.increment_value)
-        self.spinbox.bind("<Down>", self.decrement_value)
+        spinbox.bind("<Up>", self.increment_value)
+        spinbox.bind("<Down>", self.decrement_value)
 
         # Bind mouse click on spinbox arrows
-        self.spinbox.bind("<ButtonRelease-1>", self.on_spinbox_click)
+        spinbox.bind("<ButtonRelease-1>", self.on_spinbox_click)
 
         # Bind key release event to update calendar when user changes the input field
-        self.spinbox.bind("<KeyRelease>", self.on_spinbox_change)
-
-        # Toggle calendar button
-        self.toggle_button = tk.Button(self, text="Show/Hide Calendar", command=self.toggle_calendar)
-        self.toggle_button.pack(pady=10)
-
-        if Calendar:
-            self.create_calendar()
+        spinbox.bind("<KeyRelease>", self.on_spinbox_change)
+        return spinbox
 
     def bind_all_events(self):
         # Copy to clipboard with ctrl+c
@@ -51,27 +75,15 @@ class DateEntry(tk.Frame):
         # Toggle calendar widget with ctrl+shift+c
         self.bind_all("<Control-Shift-C>", lambda event: self.toggle_calendar())
 
-    def create_calendar(self):
-        # Create a frame to hold the calendar
-        self.frame = tk.Frame(self)
-        self.frame.pack(padx=20, pady=20, expand=True, fill=tk.BOTH)
-
-        # Add a calendar widget
-        self.calendar = Calendar(self.frame, selectmode='day', date_pattern='yyyy-mm-dd')
-        self.calendar.place(relwidth=0.7, relheight=0.8, anchor='n', relx=0.5)
-
-        # Bind date selection event
-        self.calendar.bind("<<CalendarSelected>>", self.on_date_select)
-
-        # Initialize calendar with the current date
-        self.update_calendar(self.spinbox.get(), '%Y-%m-%d %H:%M:%S.%f')
-
     def toggle_calendar(self, event=None):
-        if Calendar:
-            if hasattr(self, 'frame') and self.frame.winfo_ismapped():
-                self.frame.pack_forget()
-            else:
-                self.frame.pack(padx=20, pady=20, expand=True, fill=tk.BOTH)
+        if not self.calendar:
+            return
+        if self.calendar.winfo_ismapped():
+            self.frame.grid_forget()
+        else:
+            self.frame.grid(row=1, column=0)
+        self.tk_app._refresh_size()
+        return
 
     def increment_value(self, event=None):
         self.change_date(1)
@@ -112,7 +124,8 @@ class DateEntry(tk.Frame):
             split_input[part_index] = str(new_number).zfill(len(split_input[part_index]))
 
             if time:
-                new_value_str = f"{split_input[0]}-{split_input[1]}-{split_input[2]} {split_input[3]}:{split_input[4]}:{split_input[5]}.{split_input[6][:2]}"
+                new_value_str = f"{split_input[0]}-{split_input[1]}-{split_input[2]} "\
+                    f"{split_input[3]}:{split_input[4]}:{split_input[5]}.{split_input[6][:2]}"
                 string_format = '%Y-%m-%d %H:%M:%S.%f'
             else:
                 new_value_str = f"{split_input[0]}-{split_input[1]}-{split_input[2]}"
@@ -139,9 +152,9 @@ class DateEntry(tk.Frame):
         elif split_length > 3:
             if caret_pos < 14:  # hour
                 return 3
-            elif caret_pos < 17: # minute
+            elif caret_pos < 17:  # minute
                 return 4
-            elif caret_pos < 20: # second
+            elif caret_pos < 20:  # second
                 return 5
             else:               # millisecond
                 return 6
@@ -187,7 +200,7 @@ class DateEntry(tk.Frame):
         # Position the popup window in the top-left corner of the widget
         x = self.winfo_rootx()
         y = self.winfo_rooty()
-        
+
         # Position of the popup window has to be "inside" the main window or it will be focused on popup
         popup.geometry(f"400x100+{x+200}+{y-150}")
 
@@ -196,7 +209,6 @@ class DateEntry(tk.Frame):
 
         # Keep focus on the spinbox
         self.spinbox.focus_force()
-
 
     def select_all(self, event=None):
         self.spinbox.selection_range(0, tk.END)
@@ -207,31 +219,3 @@ class DateEntry(tk.Frame):
     def paste_from_clipboard(self, event=None):
         self.spinbox.delete(0, tk.END)
         self.spinbox.insert(0, self.clipboard_get())
-
-if __name__ == "__main__":
-    root = tk.Tk()
-    # Get the screen width and height
-    # This is calculating the position of the TOTAL dimentions of all screens combined
-    # How to calculate the position of the window on the current screen?
-    screen_width = root.winfo_screenwidth()
-    screen_height = root.winfo_screenheight()
-
-    print(screen_width, screen_height)
-
-    # Calculate the position to center the window
-    x = (screen_width // 2) - 400
-    y = (screen_height // 2) - 600
-
-    print(x, y)
-
-    # Set the position of the window
-    root.geometry(f"800x600+{x}+{y}")
-    # keep the main widget on top all the time
-    root.wm_attributes("-topmost", False)
-    root.wm_attributes("-topmost", True)
-    root.title("Date Editor")
-
-    date_entry = DateEntry(root)
-    date_entry.pack(expand=True, fill=tk.BOTH)
-    root.mainloop()
-