+++ mininterface/types.py
@@ -1,6 +1,7 @@
 from dataclasses import dataclass
+from datetime import date, datetime, time
 from pathlib import Path
-from typing import Any, Callable
+from typing import Any, Callable, Optional
 from typing_extensions import Self, override
 
 from .auxiliary import common_iterables
@@ -155,8 +156,83 @@ class PathTag(Tag):
                     break
 
 
-# TODO
-# @dataclass
-# class DateTag(Tag):
-#     """ TODO """
-#     pass
+@dataclass(repr=False)
+class DatetimeTag(Tag):
+    """
+    !!! warning
+        Experimental. Still in development.
+
+    Datetime is supported.
+
+    ```python3
+    from datetime import datetime
+    from dataclasses import dataclass
+    from mininterface import run
+
+    @dataclass
+    class Env:
+        my_date: datetime
+
+    m = run(Env)
+    ```
+
+    The arrows change the day (or the datetime part the keyboard caret is currently editing).
+
+    ![Datetime](asset/datetimetag_datetime.avif)
+
+    In this code, we want only the date part.
+
+    ```python3
+    from datetime import date
+    from dataclasses import dataclass
+    from mininterface import run
+
+    @dataclass
+    class Env:
+        my_date: date
+
+    m = run(Env)
+    ```
+
+    After clicking the button (or hitting `Ctrl+Shift+C`), a calendar appear.
+
+    ![Date with calendar](asset/datetimetag_date_calendar.avif)
+    """
+
+    # NOTE, document using full_precision.
+    # You may use the DatetimeTag to specify more options.
+
+    # ```python3
+    # from mininterface import run
+    # from mininterface.types import DatetimeTag
+
+    # run().form({
+    #     "my_date": DatetimeTag(time=True)
+    # })
+    # ```
+
+    # ![Time only](asset/datetime_time.avif)
+
+    # NOTE: It would be nice we might put any date format to be parsed.
+    # NOTE: The parameters are still ignored.
+
+    date: bool = False
+    """ The date part is active """
+
+    time: bool = False
+    """ The time part is active """
+
+    full_precision: Optional[bool] = None
+    """ Include full time precison, seconds, microseconds. """
+
+    def __post_init__(self):
+        super().__post_init__()
+        if self.annotation:
+            self.date = issubclass(self.annotation, date)
+            self.time = issubclass(self.annotation, time) or issubclass(self.annotation, datetime)
+        if not self.date and not self.time:
+            self.date = self.time = True
+        # NOTE: self.full_precision ...
+
+    def _make_default_value(self):
+        return datetime.now()